refactor: renamed /token to /core
Core folder now includes only the OAuth protocol
implementation.
