+++ projects/angular-simple-oidc/src/lib/core/angular-simple-oidc-core.module.ts
