+++ projects/angular-simple-oidc/src/lib/core/models.ts
