+++ projects/angular-simple-oidc/src/lib/core/token-crypto.service.ts
