+++ projects/angular-simple-oidc/src/lib/core/token-helper.service.spec.ts
