+++ projects/angular-simple-oidc/src/lib/core/token-validation.service.ts
