+++ projects/angular-simple-oidc/src/lib/core/validation-result.ts
