+++ projects/angular-simple-oidc/src/public-api.ts
@@ -9,10 +9,10 @@ export { AuthService } from './lib/auth.service';
 export { TokenStorageService } from './lib/token-storage.service';
 export { OidcDiscoveryDocClient } from './lib/discovery-document/oidc-discovery-doc-client.service';
 
-export { TokenCryptoService } from './lib/token/token-crypto.service';
-export { TokenHelperService } from './lib/token/token-helper.service';
-export { TokenUrlService } from './lib/token/token-url.service';
-export { TokenValidationService } from './lib/token/token-validation.service';
+export { TokenCryptoService } from './lib/core/token-crypto.service';
+export { TokenHelperService } from './lib/core/token-helper.service';
+export { TokenUrlService } from './lib/core/token-url.service';
+export { TokenValidationService } from './lib/core/token-validation.service';
 
 export { AuthConfig } from './lib/config/models';
 export {
@@ -22,7 +22,7 @@ export {
     LocalState,
     DecodedIdentityToken,
     TokenRequestResult
-} from './lib/token/models';
+} from './lib/core/models';
 
 export { RefreshTokenClient } from './lib/refresh-token-client.service';
-export { RefreshTokenValidationService } from './lib/token/refresh-token/refresh-token-validation.service';
+export { RefreshTokenValidationService } from './lib/core/refresh-token/refresh-token-validation.service';
